requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
joblib==1.3.2
gunicorn==21.2.0